internal/helper: Move Create/ExtractTarArchive tests in the helper pkg
