+++ internal/helper/testdata/rootfs_tarballs/ping.tar
